Make the Wall display scores
The Names column had to be sacrificed to make room for the scores
without changing the font size. This might be revisited.

The main display now works like a leader board. It displays the
scores, and is also sorted with the highest scores on top.

AND!XOR badges are displayed at the bottom with a score of N/A.

Any badge which seems to be eligible to have a trinket dispensed
based on its claimed score and claimed last level dispensed
is flagged with an exclamation point in the left column.

Numerous code formatting changes recommended by pycodestyle,
but I've resisted complying with its rule about line length.